Header refactoring: unify and de-catchall mutex module
