+++ include/jemalloc/internal/base_structs.h
@@ -2,6 +2,7 @@
 #define JEMALLOC_INTERNAL_BASE_STRUCTS_H
 
 #include "jemalloc/internal/jemalloc_internal_types.h"
+#include "jemalloc/internal/mutex.h"
 #include "jemalloc/internal/size_classes.h"
 
 /* Embedded at the beginning of every block of base-managed virtual memory. */
