+++ include/jemalloc/internal/extent_externs.h
@@ -1,8 +1,9 @@
 #ifndef JEMALLOC_INTERNAL_EXTENT_EXTERNS_H
 #define JEMALLOC_INTERNAL_EXTENT_EXTERNS_H
 
-#include "jemalloc/internal/rb.h"
+#include "jemalloc/internal/mutex.h"
 #include "jemalloc/internal/ph.h"
+#include "jemalloc/internal/rb.h"
 
 extern rtree_t			extents_rtree;
 extern const extent_hooks_t	extent_hooks_default;
