+++ include/jemalloc/internal/tcache_structs.h
@@ -3,7 +3,7 @@
 
 #include "jemalloc/internal/ql.h"
 #include "jemalloc/internal/size_classes.h"
-#include "jemalloc/internal/stats.h"
+#include "jemalloc/internal/stats_tsd.h"
 #include "jemalloc/internal/ticker.h"
 
 /*
