+++ src/extent.c
@@ -4,6 +4,7 @@
 
 #include "jemalloc/internal/assert.h"
 #include "jemalloc/internal/ph.h"
+#include "jemalloc/internal/mutex.h"
 
 /******************************************************************************/
 /* Data. */
