+++ src/rtree.c
@@ -3,6 +3,7 @@
 #include "jemalloc/internal/jemalloc_internal_includes.h"
 
 #include "jemalloc/internal/assert.h"
+#include "jemalloc/internal/mutex.h"
 
 /*
  * Only the most significant bits of keys passed to rtree_{read,write}() are
